refactor: restructure infra and platform directories for clarity
