+++ infrastructure/scripts/entrypoint.sh
