+++ internal/adapters/secondary/db/postgres/connection.go
@@ -10,7 +10,7 @@ import (
 	"github.com/lechitz/AionApi/internal/adapters/secondary/db/postgres/constants"
 
 	"github.com/lechitz/AionApi/internal/core/ports/output/logger"
-	"github.com/lechitz/AionApi/internal/infra/config"
+	"github.com/lechitz/AionApi/internal/platform/config"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
